backend-django/my_project_ia_rag_aws/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
boto3==1.42.1
botocore==1.42.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
drf-spectacular==0.27.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-aws==0.1.0
langchain-core==0.1.53
langsmith==0.1.147
lxml==6.0.2
Markdown==3.5
numpy==2.3.5
orjson==3.11.4
packaging==23.2
psycopg==3.3.1
psycopg-binary==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
tenacity==8.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.3.7
uritemplate==4.2.0
urllib3==2.5.0
